pyproject.toml
[project]
name = "menuextraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1.6",
    "openai>=2.13.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.26.6",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "streamlit>=1.52.2",
]
